♻️ refactor: refactor imports of components and images
