+++ src/App.js
@@ -1,9 +1,9 @@
+import { QueryClient, QueryClientProvider } from "react-query";
+import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
+
+import AddItem from "./pages/AddItem";
 import Home from "./pages/Home";
 import Items from "./pages/Items";
-import AddItem from "./pages/AddItem";
-
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { QueryClientProvider, QueryClient } from "react-query";
 
 const queryClient = new QueryClient();
