+++ src/components/RankedItems.js
@@ -1,8 +1,7 @@
-import { useRankedProductCountStore } from "../store/productCountStore";
 import useFetchItems from "../api/useFetchItems";
+import { favoriteIcon } from "../images";
+import { useRankedProductCountStore } from "../store/productCountStore";
 import formatNumber from "../utils/formatNumber";
-
-import favoriteIcon from "../images/ic_heart.png";
 
 export default function RankedItems() {
   // 화면 전환 시 달라지는 베스트 상품 데이터들을 전역적으로 관리하였습니다.
