+++ src/components/index.js
@@ -1,0 +1,6 @@
+export { default as AllItemsList } from "./AllItemsList";
+export { default as Navbar } from "./Navbar";
+export { default as Pagination } from "./Pagination";
+export { default as ProductForm } from "./ProductForm";
+export { default as RankedItems } from "./RankedItems";
+export { default as SortDropdown } from "./SortDropdown";
