+++ src/images/index.js
@@ -1,0 +1,9 @@
+export { default as arrowDown } from "./ic_arrow_down.png";
+export { default as arrowLeft } from "./arrow_left.png";
+export { default as arrowRight } from "./arrow_right.png";
+export { default as closeTagIcon } from "./ic_X.png";
+export { default as deletePreviewImageIcon } from "./ic_X_preview_image.png";
+export { default as favoriteIcon } from "./ic_heart.png";
+export { default as pandaLogo } from "./panda_logo.png";
+export { default as searchIcon } from "./ic_search.png";
+export { default as sortButton } from "./btn_sort.png";
