+++ src/pages/AddItem.js
@@ -1,5 +1,4 @@
-import Navbar from "../components/Navbar";
-import ProductForm from "../components/ProductForm";
+import { Navbar, ProductForm } from "../components";
 
 // 상품 등록 페이지입니다.
 export default function AddItem() {
