+++ src/pages/Home.js
@@ -1,4 +1,4 @@
-import Navbar from "../components/Navbar";
+import { Navbar } from "../components";
 
 // 추후 리액트로 병합하기 전 홈페이지입니다.
 export default function Home() {
